Version 1.0, connects to irc and generates phrases
The bot connects to irc and identifies to nickserv.
Upon getting the command !marko [N], where N is read as a number
defaulting to 1 in case of omission or failure to parse.
The irc code is a slight adaptation of the code by Don Stewart
found on the haskell wiki:
  https://wiki.haskell.org/Roll_your_own_IRC_bot
The markov chain code is based on a blog post by Aakash Japi
found here:
  http://aakashjapi.com/mimicking-writing-style-with-markov-chains